Extract editor initialization into reusable function
