+++ .github/workflows/test.yml
@@ -14,4 +14,5 @@ jobs:
           node-version: 20
           cache: npm
       - run: npm ci
+      - run: npm run lint
       - run: npm test
