+++ tsconfig.json
@@ -5,7 +5,7 @@
     "moduleResolution": "node",
     "outDir": "dist",
     "rootDir": "src",
-    "strict": false,
+    "strict": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
     "lib": ["ES2020", "DOM"]
